ENH: DICOM browser rework
User features:
- Improved DICOM import: it runs in the background (GUI is not blocked), continuous, detailed progress information is displayed; and it is about 3x faster than before
- DICOM patient/study/series can be loaded by double-clicking on the item
- Improved DICOM module GUI: optimized column widths, default sorting (based on last insertion time, study date, and series number) in DICOM browser
- DICOM database version upgrade is not enforced when updating Slicer (user can choose between upgrade or creating a new database at a new location)
- Fixed starting and stopping of DICOM listener on Windows.
- Remove orphan empty directories after deleting files from the database
- Fixed crash when query/retrieve was used from Python (where database cannot be passed to the Q/R widget as a shared pointer)

Developer features:
- Moved many DICOM features into CTK (export to file system, DICOM metadata display, etc.)
- Separated DICOM widget from logic: DICOM database is now owned by Slicer application using a shared pointer. This allows DICOM import and loading without using GUI classes (without instantiating DICOM browser widget). See example in DICOMUtils.loadSeriesByUID.
- DICOM database object is always created (even if no valid DICOM database location is specified) when the application is started. If no valid DICOM database location is specified then the database will not be open.
- Made recent activity widget selection trigget selection of the series in DICOM browser
- Fixed tests (JRC2013Vis and RSNAVisTutorial, SubjectHierarchyGenericSelfTest)
