+++ Modules/Scripted/DICOMLib/__init__.py
@@ -1,6 +1,9 @@
 from .DICOMProcesses import *
 from .DICOMExportScalarVolume import *
 from .DICOMExportScene import *
-from .DICOMWidgets import *
+from .DICOMBrowser import *
 from .DICOMPlugin import *
 from .DICOMUtils import *
+from .DICOMPluginSelector import *
+from .DICOMRecentActivityWidget import *
+from .DICOMSendDialog import *
